feat(input): improve AI prediction robustness and streaming handling
- Add optNonBlankText helper for robust null/blank JSON value extraction
- Improve streaming response handling with null-safe text delta processing
- Extract message content logic into dedicated method for reuse
- Add AppUtil for clipboard and network policy checks
- Add comprehensive unit tests for LanLlmClient, LanLlmPrompt, and AiSuggestionStripComponent
- Improve UI feedback for AI suggestion states

+++ app/src/main/java/org/fxboomk/fcitx5/android/input/candidates/CandidateItemUi.kt
@@ -63,7 +63,7 @@ class CandidateItemUi(
         @ColorInt bgColor: Int,
         @ColorInt strokeColor: Int,
         @ColorInt pressColor: Int,
-        cornerRadius: Float = ctx.dp(6f).toFloat()
+        cornerRadius: Float = ctx.dp(6f)
     ) {
         root.background = firstCandidateDrawable(
             bgColor = bgColor,
